history - limit long running shutdown handling to desktop only
