+++ src/vs/workbench/services/workingCopy/browser/workingCopyHistoryService.ts
@@ -1,0 +1,33 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IFileService } from 'vs/platform/files/common/files';
+import { IRemoteAgentService } from 'vs/workbench/services/remote/common/remoteAgentService';
+import { IEnvironmentService } from 'vs/platform/environment/common/environment';
+import { IUriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentity';
+import { ILabelService } from 'vs/platform/label/common/label';
+import { ILogService } from 'vs/platform/log/common/log';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { WorkingCopyHistoryService } from 'vs/workbench/services/workingCopy/common/workingCopyHistoryService';
+import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
+import { IWorkingCopyHistoryService } from 'vs/workbench/services/workingCopy/common/workingCopyHistory';
+
+export class BrowserWorkingCopyHistoryService extends WorkingCopyHistoryService {
+
+	constructor(
+		@IFileService fileService: IFileService,
+		@IRemoteAgentService remoteAgentService: IRemoteAgentService,
+		@IEnvironmentService environmentService: IEnvironmentService,
+		@IUriIdentityService uriIdentityService: IUriIdentityService,
+		@ILabelService labelService: ILabelService,
+		@ILogService logService: ILogService,
+		@IConfigurationService configurationService: IConfigurationService
+	) {
+		super(fileService, remoteAgentService, environmentService, uriIdentityService, labelService, logService, configurationService);
+	}
+}
+
+// Register Service
+registerSingleton(IWorkingCopyHistoryService, BrowserWorkingCopyHistoryService, true);
